refactor: moved locate method for cylinder from "oncylinder" to "locate" module
