+++ acc/geometry/locate.py
@@ -38,12 +38,11 @@ class LocateFuncFactory:
         return locateWrtSphere
 
     def onCylinder(self, cyl):
-        from . import oncylinder
         R = cyl.radius/units.length.meter
         H = cyl.height/units.length.meter
         @cuda.jit(device=True, inline=True)
         def locateWrtCylinder(x, y, z):
-            return oncylinder.cu_device_locate_wrt_cylinder(x,y,z, R, H)
+            return cu_device_locate_wrt_cylinder(x,y,z, R, H)
         return locateWrtCylinder
 
 # device functions for solid shapes
@@ -53,3 +52,17 @@ def cu_device_locate_wrt_sphere(x,y,z, R):
     if dist2>(R+epsilon)*(R+epsilon): return outside
     elif dist2<(R-epsilon)*(R-epsilon): return inside
     return onborder
+
+@cuda.jit(device=True, inline=True)
+def cu_device_locate_wrt_cylinder(x,y,z, R, H):
+    if ( abs(z)-H/2. > epsilon ) :
+        return outside
+
+    r2 = x*x+y*y
+    if r2 > (R+epsilon)*(R+epsilon):
+        return outside
+
+    if H/2.-abs(z) > epsilon and r2 < (R-epsilon)*(R-epsilon):
+        return inside
+    return onborder
+
